Added Support for VNC Kerberos Proxying to WebSocketProxy
This introduces support for doing man-in-the-middle kerberos
encryption between the proxy and the hosts.  Two config options,
vnc_krb_auth and vnc_krb_username, are used to enable the option
(additionally, SASL and GSSAPI support will have to be enabled
for QEmu's VNC).  The proxy now pretends to be a RFB server and
and RFB client.   The "client" supports sec type 20 (SASL), while
the server supports sec type 1 (None).